+++ sys/fs/unionfs/union_vnops.c
@@ -1106,7 +1106,7 @@ static int
 union_ioctl(ap)
 	struct vop_ioctl_args /* {
 		struct vnode *a_vp;
-		int  a_command;
+		u_long  a_command;
 		caddr_t  a_data;
 		int  a_fflag;
 		struct ucred *a_cred;
